requirements.txt
fastapi
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic
uvicorn
pytest
httpx
pytest-cov
pytest-asyncio
